requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
audioop-lts==0.2.2
babel==2.17.0
bitsandbytes==0.48.1
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
fastapi==0.119.1
ffmpy==0.6.3
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
peft==0.17.1
pillow==11.3.0
psutil==7.1.1
psycopg2==2.9.11
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2025.10.22
requests==2.32.5
rich==14.2.0
ruff==0.14.1
safehttpx==0.1.6
safetensors==0.6.2
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
sympy==1.14.0
tkcalendar==1.6.1
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
